refactor: rename methods which transform Maps
By the previous convention, methods which transform a Map from string to
a type, would be named for the plural of the type transform method.
Although this was short, it would fail for type names which are already
plural (Items, External Docs, Paths) and it causes collisions when there
are conversions for both Map and Array collections.

Rename methods which transform a Map to end with "Map".

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>