refactor: remove transform*Map and transform*Array
I was overly eager when adding methods and I now think that these were
poorly conceived.  The structural position is much more important than
the static type.  For example, transformSchemaProperties and
transformComponentsSchemas are likely to be much more useful than
transformSchemaMap.

The value of having methods like transformSchemaProperties is limited.
The only advantage of such a method is slightly less, slightly clearer,
and slightly faster code for inheritors (compared to modifying
.properties in transformSchema, with removing before and adding after
super.transformSchema to avoid transforming twice for performance
parity).  The disadvantage is a larger API surface to maintain and more
difficulty adding future optimizations.  This should be considered on a
case-by-case basis.

Currently transformPaths fits the threshold (also because it has a
defined type), transformSchemaProperties is on the edge, others are
farther away.  Err on the side of smaller API for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>