remove special object prototype handling
The code is currently inconsistent about how it handles objects with
prototype other than Object.prototype.  Sometimes it is preserved,
usually it is not.

Remove any special handling and always return transformed objects with
Object.prototype, favoring consistency and code-simplicity, along with
behavior which matches what would be expected from parsed JSON (which is
how the objects are commonly loaded/stored).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>