treat undefined values more consistently
Most places in the code do not call transform methods for properties
with value `undefined`, but a few use `hasOwnProperty` or `Object.keys`
and do.  Change these to consistently only call transform methods on
values which are not `undefined` to avoid making a distinction between
missing and undefined properties.

Continue copying the properties to result objects for consistency and
code simplicity.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>