fix: call transformRequestBody without content
The convention used by OpenApiTransformerBase is that transform methods
are called regardless of the validity of the object being passed (e.g.
they may be called with integers or other non-object values).  For the
same reason, transformRequestBody should be called whenever requestBody
is not undefined, to give overridden methods the opportunity to
transform the invalid value (e.g. to produce a valid one).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>