feat: create #transformPath Array
OpenApiTransformerBase#transformPath is a read-only property with an
Array of the property names traversed during the current transformation.
This can be useful for conditional transformations (e.g.  `if
(this.transformPath[0] === 'paths')` to check if a schema is in `paths`
or `components`/`definitions`) and in error reporting (so users/callers
can determine where in the document an error occurred).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>